+++ src/Numerical/HBLAS/BLAS/Internal.hs
@@ -3,13 +3,11 @@
 module Numerical.HBLAS.BLAS.Internal(
     GemmFun
     ,SymmFun
-    ,GemvFun
     ,GerFun
     ,TrsvFun
 
     ,gemmAbstraction
     ,symmAbstraction
-    ,gemvAbstraction
     ,gerAbstraction
     ,trsvAbstraction
     ) where
@@ -30,9 +28,6 @@ type GemmFun el orient s m = Transpose ->Transpose ->  el -> el  -> MDenseMatrix
 type SymmFun el orient s m = EquationSide -> MatUpLo -> el -> el -> MDenseMatrix s orient el
   -> MDenseMatrix s orient el -> MDenseMatrix s orient el -> m ()
 
-type GemvFun el orient s m = Transpose -> el -> el
-  -> MDenseMatrix s orient el -> MDenseVector s Direct el -> MDenseVector s Direct el -> m ()
-
 type GerFun el orient s m =
     el -> MDenseVector s Direct el -> MDenseVector s Direct el -> MDenseMatrix s orient el -> m ()
 
@@ -42,9 +37,6 @@ type TrsvFun el orient s m =
 
 gemmComplexity :: Integral a => a -> a -> a -> Int64
 gemmComplexity a b c = fromIntegral a * fromIntegral b *fromIntegral c  -- this will be wrong by some constant factor, albeit a small one
-
-gemvComplexity :: Integral a => a -> a -> Int64
-gemvComplexity a b = fromIntegral a * fromIntegral b
 
 
 -- this covers the ~6 cases for checking the dimensions for GEMM quite nicely
@@ -66,15 +58,6 @@ isBadSymm RightSide ax ay bx by cx cy = isBadSymmBothSide ax ay bx by cx cy
 isBadSymmBothSide :: (Ord a, Num a) => a -> a -> a -> a -> a -> a -> Bool
 isBadSymmBothSide ax ay bx by cx cy = (minimum [ax, ay, bx, by, cx, cy] <= 0)
     || not (ax == ay && bx == cx && by == cy)
-
--- / checks if the size of a matrices rows matches input vector size
--- and the  column count matchesresult vector size
-isBadGemv :: Transpose -> Int -> Int -> Int -> Int -> Bool
-isBadGemv tr ax ay bdim cdim = isBadGemvHelper (cds tr (ax,ay))
-    where
-    cds = coordSwapper
-    isBadGemvHelper (realX,realY)  =
-            minimum [realY,realX,bdim,cdim] <= 0 ||  not (realX == bdim && realY == cdim )
 
 {-
 A key design goal of this ffi is to provide *safe* throughput guarantees
@@ -157,39 +140,6 @@ symmAbstraction symmName symmSafeFFI symmUnsafeFFI constHandler = symm
                         unsafePrimToPrim $!  (if shouldCallFast cy cx ax then symmUnsafeFFI  else symmSafeFFI)
                             rawOrder rawSide rawUplo (fromIntegral cy) (fromIntegral cx)
                                 alphaPtr ap (fromIntegral astride) bp (fromIntegral bstride) betaPtr cp (fromIntegral cstride)
-
-
-{-# NOINLINE gemvAbstraction #-}
-gemvAbstraction :: (SM.Storable el, PrimMonad m)
-                => String
-                -> GemvFunFFI scale el
-                -> GemvFunFFI scale el
-                -> (el -> (scale -> m ())-> m ())
-                -> forall orient . GemvFun el orient (PrimState m) m
-gemvAbstraction gemvName gemvSafeFFI gemvUnsafeFFI constHandler = gemv
-  where
-    shouldCallFast :: Int -> Int  -> Bool
-    shouldCallFast a b = flopsThreshold >= gemvComplexity a b
-    gemv tr alpha beta
-      (MutableDenseMatrix ornta ax ay astride abuff)
-      (MutableDenseVector _ bdim bstride bbuff)
-      (MutableDenseVector _ cdim cstride cbuff)
-        | isBadGemv tr ax ay bdim cdim =  error $! "Bad dimension args to GEMV: ax ay xdim ydim: " ++ show [ax, ay, bdim, cdim]
-        | SM.overlaps abuff cbuff || SM.overlaps bbuff cbuff =
-            error $! "The read and write inputs for: " ++ gemvName ++ " overlap. This is a programmer error. Please fix."
-        | otherwise = call
-            where
-              (newx,newy) = coordSwapper tr (ax,ay)
-              call = unsafeWithPrim abuff $ \ap ->
-                     unsafeWithPrim bbuff $ \bp ->
-                     unsafeWithPrim cbuff $ \cp ->
-                     constHandler alpha $ \alphaPtr ->
-                     constHandler beta  $ \betaPtr  ->
-                       unsafePrimToPrim $! (if shouldCallFast newx newy  then gemvUnsafeFFI else gemvSafeFFI)
-                         (encodeNiceOrder ornta) (encodeFFITranspose tr)
-                         (fromIntegral newx) (fromIntegral newy) alphaPtr ap (fromIntegral astride) bp
-                         (fromIntegral bstride) betaPtr cp (fromIntegral cstride)
-
 
 {-# NOINLINE gerAbstraction #-}
 gerAbstraction :: (SM.Storable el, PrimMonad m)
